this contains the code for my clock project
In this improved version of the clock project, I started off by swapping out the two planets (symbolizing the hour hand and the center it orbits around), UFO (symbolizing the minute hand) for clipart images. This made it look a lot cleaner and nicer. Lastly, I changed the color of the alien that symbolized the seconds to match its UFO.